+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.24.0 (2025-01-15)
+
 - added: `onSeenTxCheckpoint` callback for new Seen Tx Checkpoint API.
 - added: `onTransactions` callback to support new Seen Tx Checkpoint API.